+++ CHANGELOG.md
@@ -7,8 +7,22 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.6.2] - 2023-10-19
+
+- Move shouldRefreshWithAnchoring configuration so it is possible to disable it from outside FlashList by invalidating layout
+  - https://github.com/Shopify/flash-list/pull/935
+
+## [1.6.1] - 2023-09-14
+
+- Prevent an expired layout provider from being used again
+  - https://github.com/Shopify/flash-list/pull/915
+
+## [1.6.0] - 2023-09-13
+
 - Update types to match `react-native@0.72` view types.
   - https://github.com/Shopify/flash-list/pull/890
+- Add option to clear cached layouts on update
+  - https://github.com/Shopify/flash-list/pull/910
 
 ## [1.5.0] - 2023-07-12